vicbip/packages/pipeline/requirements.txt
pandas==2.2.2
geopandas==0.14.4
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.44.0
rapidfuzz==3.9.3
python-dotenv==1.0.1
lxml==5.2.2
